+++ src/js_results.hpp
@@ -169,7 +169,7 @@ inline void setup_aliases(parser::KeyPathMapping &mapping, const realm::SharedRe
 
         for (const Property &property : it->persisted_properties) {
             const TableRef table = ObjectStore::table_for_object_type(realm->read_group(), it->name);
-            mapping.add_mapping(table, property.alias, property.name);
+            mapping.add_mapping(table, property.public_name, property.name);
         }
     }
 }