+++ src/js_schema.hpp
@@ -63,7 +63,7 @@ typename T::Object Schema<T>::dict_for_property_array(ContextType ctx, const Obj
     for (uint32_t i = 0; i < count; i++) {
         ValueType value = Object::get_property(ctx, array, i);
         Property prop = object_schema.persisted_properties[i];
-        Object::set_property(ctx, dict, !prop.alias.empty() ? prop.alias : prop.name, value);
+        Object::set_property(ctx, dict, !prop.public_name.empty() ? prop.public_name : prop.name, value);
     }
 
     return dict;
@@ -172,7 +172,7 @@ Property Schema<T>::parse_property(ContextType ctx, ValueType attributes, String
     static const String object_type_string = "objectType";
     static const String optional_string = "optional";
     static const String property_string = "property";
-    static const String alias_string = "alias";
+    static const String internal_name_string = "mapTo";
 
     Property prop;
     prop.name = std::move(property_name);
@@ -206,9 +206,10 @@ Property Schema<T>::parse_property(ContextType ctx, ValueType attributes, String
             prop.is_indexed = Value::validated_to_boolean(ctx, indexed_value);
         }
 
-        ValueType alias_value = Object::get_property(ctx, property_object, alias_string);
-        if (!Value::is_undefined(ctx, alias_value)) {
-            prop.alias = Value::validated_to_string(ctx, alias_value);
+        ValueType internal_name_value = Object::get_property(ctx, property_object, internal_name_string);
+        if (!Value::is_undefined(ctx, internal_name_value)) {
+            prop.public_name = prop.name;
+            prop.name = Value::validated_to_string(ctx, internal_name_value);
         }
     }
     else {
@@ -337,10 +338,12 @@ typename T::Object Schema<T>::object_for_object_schema(ContextType ctx, const Ob
 
     ObjectType properties = Object::create_empty(ctx);
     for (auto& property : object_schema.persisted_properties) {
-        Object::set_property(ctx, properties, property.name, object_for_property(ctx, property));
+        auto property_key = property.public_name.empty() ? property.name : property.public_name;
+        Object::set_property(ctx, properties, property_key, object_for_property(ctx, property));
     }
     for (auto& property : object_schema.computed_properties) {
-        Object::set_property(ctx, properties, property.name, object_for_property(ctx, property));
+        auto property_key = property.public_name.empty() ? property.name : property.public_name;
+        Object::set_property(ctx, properties, property_key, object_for_property(ctx, property));
     }
 
     static const String properties_string = "properties";
@@ -359,7 +362,7 @@ typename T::Object Schema<T>::object_for_property(ContextType ctx, const Propert
     ObjectType object = Object::create_empty(ctx);
 
     static const String name_string = "name";
-    Object::set_property(ctx, object, name_string, Value::from_string(ctx, property.name));
+    Object::set_property(ctx, object, name_string, Value::from_string(ctx, property.public_name.empty() ? property.name : property.public_name));
 
     static const String type_string = "type";
     if (is_array(property.type)) {
@@ -393,8 +396,8 @@ typename T::Object Schema<T>::object_for_property(ContextType ctx, const Propert
     static const String optional_string = "optional";
     Object::set_property(ctx, object, optional_string, Value::from_boolean(ctx, is_nullable(property.type)));
 
-    static const String alias_string =  "alias";
-    Object::set_property(ctx, object, alias_string, property.alias.empty() ? Value::from_undefined(ctx) : Value::from_string(ctx, property.alias));
+    static const String mapped_to_string =  "mappedTo";
+    Object::set_property(ctx, object, mapped_to_string, Value::from_string(ctx, property.name));
 
     return object;
 }