+++ tests/js/alias-tests.js
@@ -28,7 +28,7 @@ function getRealm() {
     const schemas = [{
         name: 'ObjectA',
         properties: {
-            name: { type: 'string', alias: "otherName"},
+            otherName: { type: 'string', mapTo: "name"},
             age: {type: 'int', optional: true}
         }
     }];
@@ -58,12 +58,12 @@ module.exports = {
         const Person = {
             name: 'Person',
             properties: {
-                name:     {type: 'string', alias: '_name'},
+                _name:     {type: 'string', mapTo: 'name'},
                 address:  {type: 'string', indexed: true },
                 age:      'double',
-                married:  {type: 'bool', default: false, alias: '_married'},
-                children: {type: 'list', objectType: 'Person', alias: '_children'},
-                parents:  {type: 'linkingObjects', objectType: 'Person', property: 'children', alias: '_parents'},
+                _married:  {type: 'bool', default: false, mapTo: 'married'},
+                _children: {type: 'list', objectType: 'Person', mapTo: 'children'},
+                _parents:  {type: 'linkingObjects', objectType: 'Person', property: 'children', mapTo: 'parents'},
             }
         };
 
@@ -71,14 +71,14 @@ module.exports = {
             schema: [Person]
         });
 
-        // Aliases are only set if explicitly defined in the schema
+        // Mapped properties are reported for all variants, no matter if the public_name is set or not.
         const props = realm.schema[0].properties;
-        TestCase.assertEqual(props['name'].alias, '_name');
-        TestCase.assertEqual(props['address'].alias, undefined);
-        TestCase.assertEqual(props['age'].alias, undefined);
-        TestCase.assertEqual(props['married'].alias, '_married');
-        TestCase.assertEqual(props['children'].alias, '_children');
-        TestCase.assertEqual(props['parents'].alias, '_parents');
+        TestCase.assertEqual(props['_name'].mappedTo, 'name');
+        TestCase.assertEqual(props['address'].mappedTo, 'address');
+        TestCase.assertEqual(props['age'].mappedTo, 'age');
+        TestCase.assertEqual(props['_married'].mappedTo, 'married');
+        TestCase.assertEqual(props['_children'].mappedTo, 'children');
+        TestCase.assertEqual(props['_parents'].mappedTo, 'parents');
     },
 
 
@@ -113,8 +113,7 @@ module.exports = {
         obj.age = 1;
         TestCase.assertEqual(obj.age, 1);
 
-        // If an alias is defined, it must be used when setting properties. Using the internal name
-        // doesn't work.
+        // Even if a mapped name is set, only the public name can be used when updating properties.
         obj.name = "Baz";
         TestCase.assertEqual(obj.otherName, "Bar");
     },
@@ -123,13 +122,13 @@ module.exports = {
         const realm = getRealm();
         addTestObjects(realm);
 
-        // Aliases must be used when reading properties
+        // The mapped property names cannot be used when reading properties
         let obj = realm.objects("ObjectA")[0];
         TestCase.assertEqual(obj.name, undefined);
         TestCase.assertEqual(obj.otherName, 'Foo');
         TestCase.assertEqual(obj.age, 41);
 
-        // Only aliases are visible as keys
+        // Only the Javascript property names are visible as keys, not the mapped names.
         for(var key in obj) {
             TestCase.assertFalse(key === 'name');
         }